Add content for first recipe page
I've now fleshed out the content of the first recipe with an image,
description, ingredient list, and steps for the cooking method.

I took the liberty of adding a small back link at the top of the page in
anticipation of adding further recipes. Maybe an unnecessary step?

Anyway, that completes iteration 3 of the site as described in the
project.