Complete the site by adding 2 more recipes
Following the instructions for the first recipe page I've now added a
further two recipes, bringing the total to 3. I removed the 'back' link
I had added previously since it seemed redundant on such a simple site,
and it wasn't asked for in the first place.

The main index.html was also updated to ensure proper links to all 3
recipes, presented as an unordered list in alphabetical order.

Iteration 4 is therefore complete.