Revamp index.html file
As suggested I've tweaked the index file slightly to reflect some of the
new techniques we have picked up in studying the box model.

Such as using the margin 0 auto to center elements on the page, added a
main class to be able to apply style specifically to the index instead
of the recipe pages also, changed up the colour scheme slightly.